+++ src/main/java/com/academy/mudogroupware/users/application/usecase/ApproveAcademyApplicationUseCase.java
@@ -1,9 +1,0 @@
-package com.academy.mudogroupware.users.application.usecase;
-
-import com.academy.mudogroupware.users.application.command.ApproveAcademyApplicationCommand;
-import com.academy.mudogroupware.users.application.result.ApproveAcademyApplicationResult;
-
-public interface ApproveAcademyApplicationUseCase {
-
-    ApproveAcademyApplicationResult approve(ApproveAcademyApplicationCommand command);
-}