+++ src/main/java/com/academy/mudogroupware/users/domain/model/Plan.java
@@ -1,6 +1,0 @@
-package com.academy.mudogroupware.users.domain.model;
-
-public enum Plan {
-    FREE,
-    PAID
-}